started working on implementation of the falling process

+++ src/y22d17/figure.rs
@@ -16,7 +16,7 @@ impl Figure {
     ///
     /// Panics if figure is empty
     pub fn height(&self) -> usize {
-        self.points.last().unwrap().y + 1
+        self.points.last().unwrap().y as usize + 1
     }
 
     /// Returns the width of this [`Figure`].
@@ -25,10 +25,10 @@ impl Figure {
     ///
     /// Panics if figure is empty
     pub fn width(&self) -> usize {
-        self.points.iter().map(|v| v.x).max().unwrap() + 1
+        self.points.iter().map(|v| v.x).max().unwrap() as usize + 1
     }
 
-    pub fn contains(&self, target_x: usize, target_y: usize) -> bool {
+    pub fn contains(&self, target_x: isize, target_y: isize) -> bool {
         self.points
             .iter()
             .take_while(|v| v.y <= target_y)
@@ -40,7 +40,7 @@ impl std::fmt::Debug for Figure {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         for y in 0..self.height() {
             for x in 0..self.width() {
-                if self.contains(x, y) {
+                if self.contains(x as isize, y as isize) {
                     write!(f, "#")?;
                 } else {
                     write!(f, ".")?;
@@ -52,8 +52,8 @@ impl std::fmt::Debug for Figure {
     }
 }
 
-impl From<Vec<(usize, usize)>> for Figure {
-    fn from(points: Vec<(usize, usize)>) -> Self {
+impl From<Vec<(isize, isize)>> for Figure {
+    fn from(points: Vec<(isize, isize)>) -> Self {
         Figure {
             points: points.into_iter().map(Vector::from).collect(),
         }
